requirements.txt
opentelemetry-exporter-otlp-proto-http
opentelemetry-api 
opentelemetry-sdk
opentelemetry-proto
opentelemetry-exporter-otlp
opentelemetry-instrumentation
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-semantic-conventions
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-wsgi
opentelemetry-exporter-otlp-proto-common
Flask



# pdLnK2VdycUAmHMLpKSIQwQvCfxZmuYXh6CPjMVY
